basic animation implementation

+++ Assets/source/map/VariantObjectBase.cs
@@ -10,7 +10,7 @@ namespace Illarion.Client.Map
         private float[] offsetX;
         private float[] offsetY;
 
-        public int AnimationSpeed {get;}
+        public float AnimationSpeed {get;}
         public bool IsAnimated { get { return AnimationSpeed != 0.0f; } }
 
         public int GetNextId(int id)
@@ -41,7 +41,7 @@ namespace Illarion.Client.Map
             return new float[] {offsetX[idIndex], offsetY[idIndex]};
         }
 
-        public VariantObjectBase(int[] ids, float[] offsetX, float[] offsetY, int animationSpeed, float red, float green, float blue, float alpha, float sizeVariance, int encodedLight, float height) : base(red, green, blue, alpha, sizeVariance, encodedLight, height)
+        public VariantObjectBase(int[] ids, float[] offsetX, float[] offsetY, float animationSpeed, float red, float green, float blue, float alpha, float sizeVariance, int encodedLight, float height) : base(red, green, blue, alpha, sizeVariance, encodedLight, height)
         {
             this.ids = ids;
             this.offsetX = offsetX;
